refactor(web): add possibility for own sanity types
