+++ .cspell.json
@@ -9,6 +9,7 @@
 		".git",
 		".sonarlint",
 		".vscode",
+		"**/*generated.*",
 		"**/coverage",
 		"**/dist",
 		"**/logs",
