+++ apps/web/sanity-typegen.json
@@ -1,6 +1,6 @@
 {
 	"path": "./src/lib/sanity/**/*.{ts,tsx,js,jsx}",
 	"schema": "../studio/schema.json",
-	"generates": "./src/types/sanity.types.ts",
+	"generates": "./src/types/sanity.types.generated.ts",
 	"overloadClientMethods": true
 }
